Added Inner Tracker layouts for study of impact of radius on flat Barrel + module types on Tilted Inner Tracker layouts. The following layouts were added: UNTILTED LAYOUTS:
IT 4025 : TDR geometry + 25x100 everywhere.
IT 404: radii in flat part as in IT501 + 25x100 everywhere.
IT 405: geometry as in IT404 + 50x50 everywhere.

TILTED LAYOUTS:
IT 501: 50x50 in BPIX, 25x100 elsewhere.
IT 502: geometry as IT501 + 50x50 everywhere.
IT 503: geometry as IT501 + 50x50 in 1x1 and 1x2 modules, 25x100 in 2x2 modules.

+ Color code was updated for Inner Tracker modules:
- 1x2_25x100 : green
- 2x2_25x100 : orange
- 1x1_50x50 : brown
- 1x2_50x50 : blue
- 2x2_50x50 : yellow